requirements-production.txt
boto3<1.10
Django<2.2
django-storages<1.8
psycopg2-binary<2.8
Pillow<5.4
pytz<2018.8
sentry-sdk<0.6
django-rest-swagger<2.3.0
djangorestframework<4.0.0
requests<2.20.0
beautifulsoup4<4.7.0
django-cors-headers<2.5.0
djangorestframework-camel-case<0.3.0
django-filter<2.1.0
django-phonenumber-field<2.1.0
phonenumbers<8.11.0
celery<4.3.0
redis<3.1.0